moved autocodemaker gui related files to their own folder (and namespace)
