+++ MasterMindCompetition/GUI/AutoCodemaker/AutoCodemakerGamemode.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
-namespace MasterMindCompetition.GUI {
+namespace MasterMindCompetition.GUI.AutoCodemaker {
 	class AutoCodemakerGamemode : IGamemode {
 		public void runGamemode() {
 			AutoCodemakerMenu menu = new AutoCodemakerMenu();
