+++ MasterMindCompetition/GUI/AutoCodemaker/MasterMindForm.resx
